using JS array for player matches, added positions test
