+++ public/js/browser.js
@@ -1,17 +1,15 @@
-var $ = jQuery = require('jquery');
-$.qtip = require('qtip2');
-//var dataTable = require('datatables');
-//do on every page
-$(document).ready(function() {
-    require('./tooltips')();
-    require('./formatHtml')();
-});
-//functions to call on demand
+global.tooltips = require('./tooltips');
+global.formatHtml = require('./formatHtml');
 global.generateCharts = require('./charts');
 global.matchTable = require('./matchTables');
 global.playerTables = require('./playerTables');
+global.playerMatches = require('./playerMatches');
 global.buildMap = require('./map');
 global.generateHistograms = require('./histograms');
-global.$ = $;
 global.statusHandler = require('./statusHandler');
 global.c3 = require('c3');
+global.$ = require('jquery');
+document.addEventListener('DOMContentLoaded', function() {
+    global.tooltips();
+    global.formatHtml();
+});
