+++ public/js/formatHtml.js
@@ -1,3 +1,4 @@
+var $ = jQuery = require('jquery');
 var utility = require('./utility');
 var formatSeconds = utility.formatSeconds;
 var format = utility.format;
